Сделал первое задание
